Feat: 추가 랭크페이지 순위 자동변경기능추가
