+++ package.json
@@ -12,7 +12,6 @@
     "bson": "^4.6.1",
     "crypto-browserify": "^3.12.0",
     "customize-cra": "^1.0.0",
-    "dotenv": "^16.0.0",
     "firebase": "^9.6.6",
     "js-cookie": "^3.0.1",
     "nanoid": "^3.2.0",
